renamed components and screen
