+++ composeApp/src/commonMain/kotlin/ru/llama/tool/presentation/chat_screen/ChatComponent.kt
@@ -1,0 +1,5 @@
+package ru.llama.tool.presentation.chat_screen
+
+interface ChatComponent {
+    fun onNextScreenClicked()
+}
