+++ composeApp/src/commonMain/kotlin/ru/llama/tool/presentation/root/App.kt
@@ -1,16 +1,10 @@
 package ru.llama.tool.presentation.root
 
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeContentPadding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material3.Button
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
@@ -21,22 +15,14 @@ import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import com.arkivanov.decompose.extensions.compose.stack.Children
 import com.arkivanov.decompose.extensions.compose.stack.animation.slide
 import com.arkivanov.decompose.extensions.compose.stack.animation.stackAnimation
 import com.arkivanov.decompose.extensions.compose.subscribeAsState
-import llamacppdektoptool.composeapp.generated.resources.Res
-import llamacppdektoptool.composeapp.generated.resources.compose_multiplatform
-import org.jetbrains.compose.resources.painterResource
-import ru.llama.tool.getPlatform
-import ru.llama.tool.presentation.screen_a.ScreenAComponent
-import ru.llama.tool.presentation.screen_b.SettingsContent
+import ru.llama.tool.presentation.chat_screen.ChatScreenContent
+import ru.llama.tool.presentation.setting_screen.SettingsContent
 
 private val DarkColorScheme = darkColorScheme(
     primary = Color(0xFFBB86FC),
@@ -94,31 +80,9 @@ fun App(root: IRootComponent) {
                 animation = stackAnimation(animator = slide())
             ) {
                 when (val child = it.instance) {
-                    is IRootComponent.Child.ScreenA -> ScreenAContent(child.component)
-                    is IRootComponent.Child.SettingScreenChild -> SettingsContent(child.component)
+                    is IRootComponent.Child.ChatContentChild -> ChatScreenContent(child.component)
+                    is IRootComponent.Child.SettingContentChild -> SettingsContent(child.component)
                 }
-            }
-        }
-    }
-}
-
-@Composable
-fun ScreenAContent(component: ScreenAComponent) {
-    Column(
-        modifier = Modifier
-            .safeContentPadding()
-            .fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        var showContent by remember { mutableStateOf(false) }
-        Button(onClick = { showContent = !showContent }) {
-            Text("Click me!")
-        }
-        AnimatedVisibility(showContent) {
-            val greeting = remember { getPlatform() }
-            Column(Modifier.fillMaxWidth(), horizontalAlignment = Alignment.CenterHorizontally) {
-                Image(painterResource(Res.drawable.compose_multiplatform), null)
-                Text("Compose: $greeting")
             }
         }
     }
