+++ composeApp/src/commonMain/kotlin/ru/llama/tool/presentation/root/IRootComponent.kt
@@ -2,9 +2,9 @@ package ru.llama.tool.presentation.root
 
 import com.arkivanov.decompose.router.stack.ChildStack
 import com.arkivanov.decompose.value.Value
-import ru.llama.tool.presentation.screen_a.ScreenAComponent
-import ru.llama.tool.presentation.screen_b.SettingsComponent
-import ru.llama.tool.presentation.screen_b.SettingsState
+import ru.llama.tool.presentation.chat_screen.ChatComponent
+import ru.llama.tool.presentation.setting_screen.SettingComponent
+import ru.llama.tool.presentation.setting_screen.SettingsState
 
 interface IRootComponent {
     val stack: Value<ChildStack<*, Child>>
@@ -14,7 +14,7 @@ interface IRootComponent {
     fun onSettingsTabClicked()
 
     sealed interface Child {
-        data class ScreenA(val component: ScreenAComponent) : Child
-        data class SettingScreenChild(val component: SettingsComponent) : Child
+        data class ChatContentChild(val component: ChatComponent) : Child
+        data class SettingContentChild(val component: SettingComponent) : Child
     }
 }
