+++ composeApp/src/commonMain/kotlin/ru/llama/tool/presentation/setting_screen/SettingComponent.kt
@@ -1,0 +1,8 @@
+package ru.llama.tool.presentation.setting_screen
+
+import com.arkivanov.decompose.value.Value
+
+interface SettingComponent {
+    val state: Value<SettingsState>
+    fun onEvent(event: SettingsEvent)
+}
